+++ backend/serviya/src/main/java/com/parosurvivors/serviya/metrics/application/ports/input/ServiceTagMetricsServicePort.java
@@ -11,5 +11,5 @@ import java.util.List;
  */
 public interface ServiceTagMetricsServicePort {
 
-    List<ServiceTagMetrics> getTagMetrics(int serviceId);
+    List<ServiceTagMetrics> getTagMetrics(Long serviceId);
 }
